backend/requirements.txt
fastapi[standard]
pyjwt
passlib[bcrypt]
argon2_cffi
cryptography
pydantic-settings